+++ tok/tok_test.go
@@ -50,7 +50,7 @@ func TestIntEncoding(t *testing.T) {
 	arr := []int64{a, b, c, d, 1, 2, 3, 4, -1, -2, -3, 0, 234, 10000, 123, -1543}
 	enc.ints = arr
 	for _, it := range arr {
-		encoded := encodeInt(int64(it))
+		encoded := encodeInt(it)
 		enc.tokens = append(enc.tokens, encoded)
 	}
 	sort.Sort(byEnc{enc})
